requirements.txt
numpy
opencv-python
matplotlibpip install matplotlib
